+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/tasks/buyer/BuyerVerifiesPreparedDelayedPayoutTx.java
@@ -39,7 +39,7 @@ import java.util.List;
 
 import lombok.extern.slf4j.Slf4j;
 
-import static bisq.core.trade.validation.TradeValidation.checkDelayedPayoutTxInputAmount;
+import static bisq.core.trade.validation.DelayedPayoutTxValidation.checkDelayedPayoutTxInputAmount;
 import static bisq.core.trade.validation.TradeValidation.checkLockTime;
 import static bisq.core.trade.validation.TradeValidation.checkRawTransactionInputsAreNotMalleable;
 import static bisq.core.trade.validation.TradeValidation.toVerifiedTransaction;
