+++ core/src/test/java/bisq/core/trade/validation/DelayedPayoutTxValidationTest.java
@@ -18,6 +18,10 @@
 package bisq.core.trade.validation;
 
 import bisq.core.dao.burningman.DelayedPayoutTxReceiverService;
+import bisq.core.offer.Offer;
+import bisq.core.trade.model.bisq_v1.Trade;
+
+import org.bitcoinj.core.Coin;
 
 import org.junit.jupiter.api.Test;
 
@@ -77,6 +81,50 @@ public class DelayedPayoutTxValidationTest {
     void checkBurningManSelectionHeightRejectsLocalHeightZero() {
         assertThrows(IllegalArgumentException.class,
                 () -> DelayedPayoutTxValidation.checkBurningManSelectionHeight(10, delayedPayoutTxReceiverService(0)));
+    }
+
+    @Test
+    void checkDelayedPayoutTxInputAmountAcceptsExpectedInputAmount() {
+        Offer offer = TradeValidationTestUtils.offer(true, Coin.valueOf(10_000), Coin.valueOf(12_000), Coin.valueOf(40_000));
+        Trade trade = TradeValidationTestUtils.trade(offer, Coin.valueOf(300), Coin.valueOf(20_000));
+        long expectedInputAmount = 42_300;
+
+        assertEquals(expectedInputAmount,
+                DelayedPayoutTxValidation.checkDelayedPayoutTxInputAmount(expectedInputAmount, trade));
+    }
+
+    @Test
+    void checkDelayedPayoutTxInputAmountRejectsUnexpectedInputAmount() {
+        Offer offer = TradeValidationTestUtils.offer(true, Coin.valueOf(10_000), Coin.valueOf(12_000), Coin.valueOf(40_000));
+        Trade trade = TradeValidationTestUtils.trade(offer, Coin.valueOf(300), Coin.valueOf(20_000));
+
+        assertThrows(IllegalArgumentException.class,
+                () -> DelayedPayoutTxValidation.checkDelayedPayoutTxInputAmount(42_299, trade));
+    }
+
+    @Test
+    void checkDelayedPayoutTxInputAmountRejectsZeroAndNegativeInputAmount() {
+        Offer offer = TradeValidationTestUtils.offer(true, Coin.valueOf(10_000), Coin.valueOf(12_000), Coin.valueOf(40_000));
+        Trade trade = TradeValidationTestUtils.trade(offer, Coin.valueOf(300), Coin.valueOf(20_000));
+
+        assertThrows(IllegalArgumentException.class,
+                () -> DelayedPayoutTxValidation.checkDelayedPayoutTxInputAmount(0, trade));
+        assertThrows(IllegalArgumentException.class,
+                () -> DelayedPayoutTxValidation.checkDelayedPayoutTxInputAmount(-1, trade));
+    }
+
+    @Test
+    void checkDelayedPayoutTxInputAmountRejectsNullTrade() {
+        assertThrows(NullPointerException.class,
+                () -> DelayedPayoutTxValidation.checkDelayedPayoutTxInputAmount(1, null));
+    }
+
+
+    @Test
+    void checkValueInToleranceRejectsInvalidExpectedValueAndFactor() {
+        assertThrows(IllegalArgumentException.class, () -> TradeValidationUtils.checkValueInTolerance(1, 0, 1));
+        assertThrows(IllegalArgumentException.class, () -> TradeValidationUtils.checkValueInTolerance(1, -1, 1));
+        assertThrows(IllegalArgumentException.class, () -> TradeValidationUtils.checkValueInTolerance(1, 1, 0.99));
     }
 
     private static DelayedPayoutTxReceiverService delayedPayoutTxReceiverService(int burningManSelectionHeight) {
